+++ src/components/index.js
@@ -1,0 +1,8 @@
+export { default as Footer } from "./footer"
+export { default as Header } from "./header"
+export { default as Layout } from "./layout"
+export { default as Menu } from "./menu"
+export { default as ExternalLink } from "./external-link"
+export { default as PostListItem } from "./post-list-item"
+export { default as PostLink } from "./post-link"
+export { default as SEO } from "./seo"
