+++ src/pages/404.js
@@ -1,7 +1,5 @@
 import React from "react"
-
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import { Layout, SEO } from "../components"
 
 const NotFoundPage = () => (
   <Layout>
