+++ src/pages/projects.js
@@ -1,7 +1,6 @@
 import React from "react"
 import styled from "styled-components"
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import { Layout, SEO } from "../components"
 
 const projects = [
   {
