+++ src/templates/post.js
@@ -1,7 +1,6 @@
 import React from "react"
 import { graphql } from "gatsby"
-import Layout from '../components/layout';
-import SEO from "../components/seo"
+import { Layout, SEO } from "../components"
 
 export default function Template({
   data, // this prop will be injected by the GraphQL query below.
